+++ client/package.json
@@ -1,7 +1,7 @@
 {
   "name": "mytinydc-utdon-client",
   "private": true,
-  "version": "1.8.0",
+  "version": "1.9.0",
   "description": "Application for tracking obsolete FOSS applications - UI",
   "type": "module",
   "scripts": {
